requirements.txt
# AI Study Buddy - Dependencies

# Install all dependencies with: pip install -r requirements.txt

# Core Streamlit app
streamlit>=1.30.0

# LangChain and related packages
langchain>=1.0.0
langchain-community>=0.3.0
langchain-core>=1.0.0
langchain-ollama>=1.0.0
langchain-text-splitters>=1.0.0
langchainhub>=0.1.0

# Ollama (local LLM)
ollama>=0.1.0

# Vector database
faiss-cpu>=1.7.0

# Document handling
pypdf>=4.0.0
unstructured[docx]>=0.12.0
pikepdf>=8.0.0
python-dotenv>=1.0.0

# Additional utilities
requests>=2.31.0
